src/video_redactor/pyproject.toml
[tool.poetry]
name = "video-redactor"
version = "0.1.0"
description = "Сервис video_redactor проекта vidos. Отвечает за конструирование видео виджета."
authors = ["eslupmi101 <ajsen-andreev@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "3.2.23"
djangorestframework = "^3.14.0"
django-rest-framework-mongoengine = "^3.4.1"
mongoengine = "^0.27.0"
pymongo = "^4.6.1"
python-dotenv = "^1.0.0"
drf-yasg = "^1.21.7"
django-cors-headers = "^4.3.1"
requests = "^2.31.0"
django-filter = "^23.5"
django-mongoengine = "^0.5.6"
gunicorn = "^21.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
